requirements.txt
openai>=1.3.0
python-dotenv>=1.0.1
streamlit>=1.28.0
httpx>=0.24.1
fpdf2>=2.5.1
python-docx>=0.8.11
markitdown[pdf,docx]>=0.1.0
pydantic>=2.0.0
python-magic>=0.4.27
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
python-magic-bin>=0.4.14; sys_platform == 'win32'
